backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
ldap3
python-multipart
pyjwt
passlib[bcrypt]
pytest
httpx
python-dotenv
